@@ -1,74 +1,118 @@
-//Biblioteca 
-import createError from "http-errors";
-//El framework express
-//var express = require('express');
-import express from "express";
-//Bibliooteca del nucleo de node que sirve para
-//administrar rutas
-//var path = require('path');
-import path from "path";
-//Biblioteca externa que sirve para administrar cookies
-import cookieParser from "cookie-parser";
-//var cookieParser = require('cookie-parser');
-//Biblioteca que registra en consola
-//solicitudes del cliente
-import logger from "morgan";
+// Biblioteca de 3ros para manejar errores http
+// ES5: var createError = require('http-errors');
+// ES6 👇
+import createError from 'http-errors';
+// El framework express
+import express from 'express';
+// Biblioteca del nucleo de node que sirve para
+// administrar rutas
+import path from 'path';
+// Biblioteca externa que sirve para administrar
+// cookies
+import cookieParser from 'cookie-parser';
+// Biblioteca que registra en consola
+// solicitudes del cliente
+import morgan from 'morgan';
+import webpack from 'webpack';
 
-//Definición de rutas
-//var indexRouter = require('./routes/index');
-import indexRouter from "./routes/index";
-//var usersRouter = require('./routes/users');
-import usersRouter from "./routes/index";
+import webpackDevMiddleware from 'webpack-dev-middleware';
+import WebpackHotMiddleware from 'webpack-hot-middleware';
 
-//var app = express();
+// Logger de la aplicación
+import logger from './config/winston';
+import debug from './services/debugLogger';
+
+// Importando Webbpack middleware
+import webpackConfig from '../webpack.dev.config';
+
+// Definición de rutas
+import indexRouter from './routes/index';
+import usersRouter from './routes/users';
+
+// Recuperar el modo de ejecución de la app
+const nodeEnv = process.env.NODE_ENV || 'development';
+
+// Creando una instancia de express
 const app = express();
 
+// Inclusion del webpack middleware
+if (nodeEnv === 'development') {
+  debug('✒ Ejecutando en modo de desarrollo 👨‍💻');
+  // Configurando webpack en modo de desarrollo
+  webpackConfig.mode = 'development';
+  // Configurar la ruta del HMR (Hot Module Replacement)
+  // 👉 "reload=true" -> Habilita la recarga automatica cuando un archivo
+  // js cambia
+  // 👉 "timeout=1000" -> Establece el timpo de refresco de la pagina
+  webpackConfig.entry = [
+    'webpack-hot-middleware/client?reload=true&timeout=1000',
+    webpackConfig.entry,
+  ];
+  // Agregando el plugin a la configuracion
+  webpackConfig.plugins.push(new webpack.HotModuleReplacementPlugin());
+  // Crear el empaquetado con webpack
+  const bundler = webpack(webpackConfig);
+  // Registro el middleware en express
+  app.use(
+    webpackDevMiddleware(bundler, {
+      publicPath: webpackConfig.output.publicPath,
+    })
+  );
+  // Registrando el HMR Middleware
+  app.use(WebpackHotMiddleware(bundler));
+} else {
+  debug('✒ Ejecutando en modo de producción 🏭');
+}
+
 // view engine setup
-//view engine setup
-//Configura el motor de plantillas
-//1. Establecer donde estarán las plantillas
-///vistas -> Views
-//app.set("<nombre de la var>", <valor>)
+// Configura el motor de plantillas
+// 1. Establecer donde estarán las plantillas
+// (Vistas -> Views)
+// app.set("<nombre de la var>", <valor>)
 app.set('views', path.join(__dirname, 'views'));
-console.log(`☢️ El valor de __dinrma es: ${__dirname}`)
-console.log(`⚠️ El valor del join es: ${path.join(__dirname, 'views')}}`)
-//JOIN TIENE LA INTELIGENCIA PARA SABER EN QUE ENTORNO ESTA Y SELECCIONA EL CARACTER CON EL QUE SELECCIONA LA RUTA ("/ O backslash")
-//Establezco que motor precargado usare
+// Establezco que motor precargado usare
 app.set('view engine', 'hbs');
 
-//[Cliente] => {get+url+cliente+...} REQUEST
-//[Cliente] => req => 💻 => req (req,res)=>{} [ middleware 1 ] => req => [ middleware 2 ] => res
-//re == {html + dns cliente}
-//[Cliente] <== <== <== <== <== <== <== 
-//Middleware son objetos que reciben req y res next invoca al siguiente Middleware
+// Establezco Middelware
+app.use(morgan('dev', { stream: logger.stream }));
+// Middleware para parsear a json la peticion
+app.use(express.json());
+// Decodificar la url
+app.use(express.urlencoded({ extended: false }));
+// Parsear cookies
+app.use(cookieParser());
+// Servidor de archivos estáticos
+app.use(express.static(path.join(__dirname, '..', 'public')));
 
-//Establezo Middelware
-app.use(logger('dev'));//logear
-app.use(express.json());//Middleware para parsear a json
-app.use(express.urlencoded({ extended: false }));// decode la URL
-app.use(cookieParser());//trabajar con cookies - parsear cookies, se guardan en el navegador
-app.use(express.static(path.join(__dirname, 'public')));//Servidor de archivos estaticos EN CARPETA public se puede servir todo de manera estatica
-
-
-//Registro Rutas
-app.use('/', indexRouter); 
+// Registro Rutas
+app.use('/', indexRouter);
+app.use('/index', indexRouter);
 app.use('/users', usersRouter);
 
 // catch 404 and forward to error handler
-app.use(function(req, res, next) {
+app.use((req, res, next) => {
+  logger.error(
+    `404 Page Not Found - ${req.originalUrl} - Method: ${req.method}`
+  );
   next(createError(404));
 });
 
 // error handler
-app.use((err, req, res, next) => {
+app.use((err, req, res) => {
   // set locals, only providing error in development
   res.locals.message = err.message;
   res.locals.error = req.app.get('env') === 'development' ? err : {};
+
+  // Registrando mensaje de error
+  logger.error(`${err.status || 500} - ${err.message}`);
 
   // render the error page
   res.status(err.status || 500);
   res.render('error');
 });
 
-//module.exports = app;
+// Exportando la instancia del server "app"
+// ES5 👇
+// module.exports = app;
+// ES6 👇
 export default app;
