@@ -1,12 +1,13 @@
-var express = require('express');
-var router = express.Router();
+const express = require('express');
+
+const router = express.Router();
 
 /* GET home page. */
-router.get('/', function(req, res, next) {
-  //View model - al objeto (title: express)
-  res.render('index', { 
+router.get('/', (req, res) => {
+  // View model - al objeto (title: express)
+  res.render('index', {
     title: 'Express',
-    author: 'Alfred Saucedo'
+    author: 'Alfred Saucedo',
   });
 });
 
